+++ pychron/experiment/tasks/experiment_editor.py
@@ -34,6 +34,9 @@ from pychron.core.helpers.filetools import add_extension
 class ExperimentEditorHandler(TabularEditorHandler):
     refresh_name = 'refresh_table_needed'
 
+    def select_unknowns(self, info, obj):
+        obj.select_unknowns()
+
     def select_same(self, info, obj):
         obj.select_same()